Merge patch-node_gyp
I'm not sure if this patch relates to building on OpenBSD generally,
or if it's only needed within the OpenBSD Ports environment.
(Nearly all of the remaining patches I have not yet merged fall into
the latter category.)

If it turns out that this patch is specific to building in OpenBSD
Ports, I may revert it later.