adapt patch-src_env_cc for use outside OpenBSD Ports (UGH)
The original patch-src_env_cc hardcodes node's process path at
build time. I guess the equivalent outside of the OpenBSD Ports
build process would be to hardcode some path, like (as this does)
`/usr/local/bin/node`.

This is janky, for sure, but process.execPath is totally busted
without this patch, and I'm hoping that if I can make process.execPath
(seem to) work, then maybe I'll be able to avoid an even jankier
patch for yarn...

(On the flip side, if this fails to fix anything, maybe I'll
revert it later.)